+++ src/components/common/StatusIndicator.tsx
@@ -1,4 +1,4 @@
-import type { Status } from '@/api/types';
+import type { Status } from '@/apis/types';
 import theme from '@/styles/theme/emotionTheme';
 
 import type { ComponentPropsWithoutRef } from 'react';
