+++ src/components/pages/main/Main.tsx
@@ -1,6 +1,6 @@
 import { Flex, HStack, VStack, Link } from '@chakra-ui/react';
 
-import summoners from '@/api/mocks/summoners';
+import summoners from '@/apis/mocks/summoners';
 import Logo from '@/assets/images/logo.svg';
 import Select from '@/components/common/select/Select';
 import type { SelectOption } from '@/components/common/select/useSelect';
